reimpl: move the new tests into their respective test files
The coverage_*_test.go files are gone: public-API tests joined
unmarshaler_test.go and marshaler_test.go, and the internal ones now
live in test files named after the source file they exercise
(decode_test.go, strict_test.go, unstable/{kind,ast}_test.go,
internal/tracker/key_test.go). No test was modified.
